feat(admin): consolidate tenant management
